#20 split core logic into modular files
 Moved into yank, delete, edit, finders, and helpers into separate
 modules.

 Closes #20
