+++ lua/ghostnotes/note_operations/getters.lua
@@ -1,0 +1,10 @@
+local M = {}
+function M.get_note_headline(note)
+	local first_line = note.text:match("([^\n]+)") or note.text
+	if note.text:find("\n") then
+		return first_line .. " [..]"
+	end
+	return first_line
+end
+
+return M
